+++ sphere.cuh
@@ -12,7 +12,7 @@
 
 class sphere: public hittable {
 public:
-    sphere(point3 _center, double _radius, shared_ptr<material> _material)
+    sphere(point3 _center, float _radius, shared_ptr<material> _material)
     : center(_center), radius(_radius), mat(_material) {}
 
     bool hit(const ray& r, interval ray_t, hit_record& rec) const override {
@@ -42,7 +42,7 @@ public:
     }
 private:
     point3 center;
-    double radius;
+    float radius;
     shared_ptr<material> mat;
 };
